requirements.txt

Django==4.1.7
djangorestframework==3.14.0
Pillow==7.0.0
djangorestframework-simplejwt-5.2.2\

requests==2.27.1
